Initial commit of a tiny computer algebra system.
To help support multidimensional calculators, we have a tiny CAS that allows you
to build up an equation with variables, change those variables at any time,
solve the equation, and calculate a concrete value for an expression as long
there are no unknown variables in the expression. The solver only works only on
a subset of simple equations which must be one equation with one unknown and an
equality relationship.