Add TotalDoseCalc.
This is a rebuilt version of the existing 'ingestion' form. The form UI uses ng-bootstrap and the new components. We take a major departure by not using angular/forms at all, that API mostly seems to get in the way.

On construction, this builds up an equation 'concen' * 'intake' * 'substanceDensity' * 'molarMass' / ('solutionDensity' * 'bodyWeight' * 'dose') and asks the algebra system to solve those equations and save them as Terms. When computing an answer, we load the values from the <input> elements into the Variable objects and call getValue() on the appropriate Term to get its result. Adjust for units and display to the user. To handle optional rows, any row not displayed has a value of '1 unitless, which works because these are all multiplies and divides.

Calculations are performed whenever an input box loses focus. We check to see if all but one input box is filled in and if so, that's our output box and we put the computation there.

There's a fascinating corner case with this logic: you can add 2 optional rows, fill in the 4 mandatory rows, then remove 2 optional rows. Now all 4 mandatory rows are filled in and none of them is the output.

We use the oninput event to forbid entry of invalid values, including negatives and scientific notation. We never output scientific notation either. Some browsers (such as Chrome) show a "spinner" widget on input type="number" so we use text type and simply check it ourselves.