Display calculation equation.
When no computation is being performed, we show D = ... even if there is a value provided for D. We switch when we calculate. We also change the equation when units change to add or remove variables.

A new equation visitor is added to serialize the equation down to HTML.

Now that equations are user-visible, perform some more simplifications on them. Multiplication by unitless 1 is simplified away, and (a*b)^n becomes a^n * b^n.

Change how reciprocal values are handled, now they have their own independent variables in the equation. This ensures we display multiplication and division correctly in the user-visible equation, instead of silently reciprocating things on input and output.

Move solutionDensityVar to the numerator and the new solutionDensityRecipVar is on the denominator. This means that we flip the meaning of solutionDensityRecip since it used to be true when on the numerator.