Give the safedose logo a content addressable storage name.
The 20-character file name extension is computed with "sha384sum assets/safedose.svg | cut -b1-20", or the first 80 bits of the sha384 as hexits. Any changes to the logo will require a change in the file name, but that change will need to be performed manually. The web server knows to vend files following this naming convention with Cache-control: immutable.