Add humancalc tutorial.
The PDF is named with content addressible storage so that the web server vends it with Cache-Control: immutable. If you edit the file, you must also rename it. The hash section is the first 20 hexits of the sha384 hash. You can use "sha384sum file.pdf | cut -b1-20" to recompute it.