Fix bug where value displayed and value in model could get out of sync.
To repro the bug, fill in "1.23" then come back and put the cursor after the '.' and type a '/'. It won't appear on the screen, but it gets added to the value in the model, which in turn causes it to be parsed as '1'. Noticed because "1./23" appears in the calculation log.